support_ticket_classification_system/pyproject.toml
[project]
name = "support-ticket-classification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langgraph>=1.1.10",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
